Add logs to services on backend for future frontend use and display.

+++ backend/src/main/java/pt/ua/deti/uasmartsignage/services/MonitorService.java
@@ -1,15 +1,19 @@
 package pt.ua.deti.uasmartsignage.services;
 
 import pt.ua.deti.uasmartsignage.models.MonitorGroup;
-import pt.ua.deti.uasmartsignage.models.Rule;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import pt.ua.deti.uasmartsignage.repositories.MonitorRepository;
+import pt.ua.deti.uasmartsignage.enums.Log;
+import pt.ua.deti.uasmartsignage.enums.Severity;
 import pt.ua.deti.uasmartsignage.models.Monitor;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -17,25 +21,34 @@ public class MonitorService {
 
     private final MonitorRepository monitorRepository;
     private final MonitorGroupService monitorGroupService;
-    private final RuleService ruleService;
     private final LogsService logsService;
+
+    private static final Logger logger = LoggerFactory.getLogger(MonitorService.class);
+    private final String source = this.getClass().getSimpleName();
 
     /**
      * Retrieves and returns a Monitor by its ID.
      *
      * @param id The ID of the Monitor to retrieve.
-     * @return The Monitor with the specified ID.
+     * @return The Monitor with the specified ID or empty if not found.
      */
-    public Monitor getMonitorById(Long id) {
+    public Optional<Monitor> getMonitorById(Long id) {
+        logger.info("Retrieving monitor with ID: {}", id);
+        Optional<Monitor> monitor = monitorRepository.findById(id);
 
-        Monitor monitor = monitorRepository.findById(id).orElse(null);
+        String operation = "getMonitorById";
+        String description = "Retrieved monitor with ID: " + id;
 
-        if (monitor == null) {
-            return null;
+        if (monitor.isEmpty()){
+            description = Log.OBJECTNOTFOUND.format(id);
+            logsService.addLogEntry(Severity.WARNING, source, operation, description, logger);
+            return Optional.empty();
         }
 
-        boolean online = logsService.keepAliveIn10min(monitor);
-        monitor.setOnline(online);
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
+
+        boolean online = logsService.keepAliveIn10min(monitor.get());
+        monitor.get().setOnline(online);
 
         return monitor;
     }
@@ -47,17 +60,38 @@ public class MonitorService {
      * @return The saved Monitor.
      */
     public Monitor saveMonitor(Monitor monitor) {
+        String operation = "saveMonitor";
+        String description = "Created monitor with name: " + monitor.getName();
+
         monitor.setId(null);
-        return monitorRepository.save(monitor);
+        Monitor savedMonitor = monitorRepository.save(monitor);
+
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
+
+        return savedMonitor;
     }
 
     /**
      * Deletes a Monitor from the database.
      *
      * @param id The ID of the Monitor to delete.
+     * @return Boolean with true if monitor was deleted or false if it failed.
      */
-    public void deleteMonitor(Long id) {
+    public Boolean deleteMonitor(Long id) {
+        Optional<Monitor> monitor = getMonitorById(id);
+
+        String operation = "deleteMonitor";
+        String description = "Deleted monitor with ID: " + id;
+
+        if (monitor.isEmpty()){
+            description = Log.OBJECTNOTFOUND.format(id);
+            logsService.addLogEntry(Severity.ERROR, source, operation, description, logger);
+            return false;
+        }
+        
         monitorRepository.deleteById(id);
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
+        return true;
     }
     
     /**
@@ -68,13 +102,15 @@ public class MonitorService {
      * @return The updated Monitor.
      */
     public Monitor updateMonitor(Long id, Monitor monitor) {
+        String operation = "updateMonitor";
+        String description = "Updated monitor with ID: " + id + "; And name: " + monitor.getName();
 
-        Monitor oldMonitor = getMonitorById(id);
+        Monitor oldMonitor = getMonitorById(id).get();
         MonitorGroup oldGroup = oldMonitor.getGroup();
 
         MonitorGroup group = null;
         if(monitor.getGroup().getId() != null){
-            group = monitorGroupService.getGroupById(monitor.getGroup().getId());
+            group = monitorGroupService.getGroupById(monitor.getGroup().getId()).get();
         }
 
         if(group == null){
@@ -91,6 +127,7 @@ public class MonitorService {
             monitorGroupService.deleteGroupById(oldGroup.getId());
         }
 
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
         return updatedMonitor;
     }
 
@@ -99,10 +136,14 @@ public class MonitorService {
      *
      * @return A list of all Monitors.
      */
-    public Monitor updatePending(Long id,boolean pending){
+    public Monitor updatePending(Long id, boolean pending){
+        String operation = "updatePending";
+        String description = "Updated monitor with ID: " + id + "; To pending: " + pending;
 
         Monitor monitorById = monitorRepository.getReferenceById(id);
         monitorById.setPending(pending);
+
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
 
         return monitorRepository.save(monitorById);
     }
@@ -114,6 +155,8 @@ public class MonitorService {
      * @return A list of all Monitors.
      */
     public List<Monitor> getAllMonitorsByPending(boolean pending) {
+        String operation = "getAllMonitorsByPending";
+        String description = "Retrieved all monitors by pending: " + pending;
 
         List<Monitor> monitors = monitorRepository.findByPending(pending);
 
@@ -121,6 +164,9 @@ public class MonitorService {
             boolean online = logsService.keepAliveIn10min(monitor);
             monitor.setOnline(online);
         }
+
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
+
         return monitors;
     }
 
@@ -133,7 +179,7 @@ public class MonitorService {
      */
     public List<Monitor> getAllMonitorsByPendingAndOnline(boolean pending, boolean onlineStatus) {
         String operation = "getAllMonitorsByPendingAndOnline";
-        String description = "Getting all monitors by pending " + pending + " and by status " + onlineStatus;
+        String description = "Retrieved all monitors by pending " + pending + " and by status " + onlineStatus;
 
         List<Monitor> monitors = monitorRepository.findByPendingAndOnline(pending, onlineStatus);
 
@@ -142,12 +188,8 @@ public class MonitorService {
             monitor.setOnline(online);
         }
 
-        // if (!logsService.addBackendLog(Severity.INFO, source, operation, description)) {
-        //     logger.error(ADDLOGERROR);
-        // }
-        // else {
-        //     logger.info(ADDLOGSUCCESS, description);
-        // }
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
+
         return monitors;
     }
 
@@ -158,7 +200,7 @@ public class MonitorService {
      */
     public List<Monitor> getMonitorsByGroupAndOnline(long groupId, boolean onlineStatus) {
         String operation = "getMonitorsByGroupAndOnline";
-        String description = "Getting all monitors by group " + groupId + " and by status " + onlineStatus;
+        String description = "Retrieved all monitors by group ID " + groupId + " and by status " + onlineStatus;
 
         List<Monitor> monitors = monitorRepository.findByPendingAndGroup_IdAndOnline(false, groupId, onlineStatus);
 
@@ -167,13 +209,7 @@ public class MonitorService {
             monitor.setOnline(online);
         }
 
-
-        // if (!logsService.addBackendLog(Severity.INFO, source, operation, description)) {
-        //     logger.error(ADDLOGERROR);
-        // }
-        // else {
-        //     logger.info(ADDLOGSUCCESS, description);
-        // }
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
         
         return monitors;
     }
@@ -184,7 +220,9 @@ public class MonitorService {
      * @return A list of all Monitors.
      */
     public List<Monitor> getMonitorsByGroup(long groupId) {
-
+        String operation = "getMonitorsByGroup";
+        String description = "Retrieved all monitors by group ID " + groupId;
+        
         List<Monitor> monitors = monitorRepository.findByPendingAndGroup_Id(false,groupId);
 
         for (Monitor monitor : monitors) {
@@ -192,6 +230,8 @@ public class MonitorService {
             monitor.setOnline(online);
         }
         
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
+
         return monitors;
     }
 
@@ -201,6 +241,11 @@ public class MonitorService {
      * @return A list of all Monitors.
     */
     public Monitor getMonitorByUUID(String uuid) {
+        String operation = "getMonitorByUUID";
+        String description = "Retrieved monitor with UUID " + uuid;
+
+        logsService.addLogEntry(Severity.INFO, source, operation, description, logger);
+
         return monitorRepository.findByUuid(uuid);
     }
